backend/exam/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
PyPDF2==3.0.1
google-generativeai==0.8.0 # Or latest stable version
numpy==1.26.4
python-dotenv==1.0.1